moved update database to update_from_api
